update user level access

+++ execution-project/app/Views/product.php
@@ -15,7 +15,9 @@
       <tbody>
       <tbody>
         <h1>Our Ice Cream Product</h1>
-        <a href="/product/add" class="btn btn-dark">Add data</a></td>
+        <?php if (session()->get('level') === '2') : ?>
+          <a href="/product/add" class="btn btn-dark">Add data</a></td>
+        <?php endif ?> 
         <?php foreach($iceCream as $row) : ?>
         <tr>
             <td><?= $row->id_iceCream ?></td>
@@ -23,8 +25,12 @@
             <td><?= $row->topping ?></td>
             <td><?= $row->tipe_harga ?></td>
             <td><?= $row->harga ?></td>
-            <td><a href="/product/edit/<?= $row->id_iceCream ?>" class="btn btn-dark">edit</a></td>
-            <td><a href="/product/delete/<?= $row->id_iceCream ?>" class="btn btn-dark">delete</a></td>
+            <?php if (session()->get('level') === '2') : ?>
+              <td><a href="/product/edit/<?= $row->id_iceCream ?>" class="btn btn-dark">edit</a></td>
+              <td><a href="/product/delete/<?= $row->id_iceCream ?>" class="btn btn-dark">delete</a></td>
+            <?php else : ?>
+              <td><a href="/order/<?= $row->id_iceCream ?>" class="btn btn-dark">Order Now</a></td>
+            <?php endif ?> 
         </tr>
         <?php endforeach; ?>
       </tbody>
